test(server): refactor test cases in common files
