+++ packages/twenty-shared/src/testing/types/EachTestingContext.type.ts
@@ -1,4 +1,7 @@
 export type EachTestingContext<T> = {
   title: string;
   context: T;
+  // TODO implem
+  skip?: boolean
+  only?: boolean
 };
