+++ packages/twenty-shared/src/testing/types/SuccessfulAndFailingTestCases.ts
@@ -1,0 +1,6 @@
+import { EachTestingContext } from '@/testing/types/EachTestingContext.type';
+
+export type SuccessfulAndFailingTestCases<T> = {
+  successful: EachTestingContext<T>[];
+  failing: EachTestingContext<T>[];
+};
